Improve logging output
If the user started processing the chain at a point close to the end,
the messages made it look like the synchronisation was already almost
finished ("Synchronising (99.96%)"), risking confusing the user.
Thats indeed the relative point on the chain that is currently being
processed and has nothing to do with an ETA.

This patch reworks the message to clear the meaning of that percentage.
E.g.  "Synchronising. Current slot 57848748 out of 57901784 (99.91%)"